requirements.txt
discord.py
dataclasses
